+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/graph/SecondPassData.java
@@ -1,0 +1,9 @@
+package com.jpexs.decompiler.graph;
+
+/**
+ *
+ * @author JPEXS
+ */
+public abstract class SecondPassData {
+
+}